+++ libs/core/button/button.component.spec.ts
@@ -103,3 +103,54 @@ describe('ButtonComponent – Disabled', () => {
         expect(cssClass).toContain('is-disabled');
     });
 });
+
+describe('ButtonComponent - AriaLabel Tests', () => {
+    let fixture: ComponentFixture<TestComponent>;
+    let debugElement: DebugElement;
+    let component: DebugElement;
+    let componentInstance: ButtonComponent;
+
+    beforeEach(waitForAsync(() => {
+        TestBed.configureTestingModule({
+            imports: [ButtonModule, TestComponent]
+        }).compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(TestComponent);
+        debugElement = fixture.debugElement;
+        fixture.detectChanges();
+        component = debugElement.query(By.directive(ButtonComponent));
+        componentInstance = component.injector.get(ButtonComponent);
+    });
+
+    it('should return input ariaLabel if provided', () => {
+        componentInstance.ariaLabel = 'Custom Aria Label';
+        expect(componentInstance.buttonArialabel).toBe('Custom Aria Label');
+    });
+
+    it('should return element native aria-label attribute if no input ariaLabel provided', () => {
+        const nativeElement: HTMLElement = componentInstance.elementRef.nativeElement;
+        nativeElement.setAttribute('aria-label', 'Native Aria Label');
+        expect(componentInstance.buttonArialabel).toBe('Native Aria Label');
+    });
+
+    it('should return label if fdType is special and label is provided', () => {
+        componentInstance.fdType = 'emphasized';
+        componentInstance.label = 'Button Label';
+        expect(componentInstance.buttonArialabel).toBe('Button Label');
+    });
+
+    it('should return glyph with hyphen replaced as spaces if fdType is special and glyph is provided', () => {
+        componentInstance.fdType = 'emphasized';
+        // Ensure label is undefined to test glyph transformation.
+        componentInstance.label = undefined;
+        componentInstance.glyph = 'icon-glyph-name';
+        expect(componentInstance.buttonArialabel).toBe('icon glyph name');
+    });
+
+    it('should return null if no conditions are met', () => {
+        componentInstance.fdType = 'standard';
+        expect(componentInstance.buttonArialabel).toBeNull();
+    });
+});